requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.26
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
